# skill-registry/internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: SkillForge Registry API
  version: 1.0.0
  description: |
    The SkillForge Registry API for publishing, discovering, and managing AI skills,
    agents, flows, prompts, tools, and bundles.

    ## Authentication

    Public read endpoints (browsing, downloading public artifacts) require no authentication.
    Write operations and token management require a Bearer token:

    ```
    Authorization: Bearer <your-token>
    ```

    Obtain a token via `POST /api/v1/auth/login` or create a long-lived API token in the
    web UI under **Account → Tokens**.

    ## Scopes

    | Scope | Description |
    |-------|-------------|
    | `read` | Read artifacts and metadata |
    | `write` | Publish and manage artifacts |
    | `delete` | Delete artifact versions |
    | `token:create` | Create API tokens |
    | `token:revoke` | Revoke API tokens |
    | `namespace:admin` | Manage namespace members |
    | `admin` | Full access including audit log |
servers:
  - url: /api/v1
    description: Registry API v1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API token obtained via POST /auth/login or the web UI
...
  schemas:
...
    ValidationResult:
      type: object
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
              file:
                type: string
              message:
                type: string
        warnings:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
              file:
                type: string
              message:
                type: string
paths:
...
  /validate:
    post:
      tags: [Skills]
      summary: Validate package
      description: |
        Validate a skill package archive without publishing it. Returns errors and warnings.

        ```bash
        skforge validate my-skill.tgz
        ```
      operationId: validate
      security:
        - bearerAuth: [write]
      requestBody:
        required: true
        content:
          application/gzip:
            schema:
              type: string
              format: binary
          application/zip:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationResult"

  # ── Auth ──────────────────────────────────────────────────────────────────
...